requirements.txt
asgiref
crispy-bootstrap5
distlib
dj-database-url
Django
django-crispy-forms
django-environ
django-filter
django-phonenumber-field
djangorestframework
filelock
Markdown
MouseInfo
phonenumbers
pillow
platformdirs
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
sqlparse
typing_extensions
tzdata
virtualenv